+++ charts/sonarqube/CHANGELOG.md
@@ -4,6 +4,8 @@ All changes to this chart will be documented in this file.
 ## [2025.6.0]
 * Update Chart's version to 2025.6.0
 * Upgrade SonarQube Community Build to 25.10.0.114319
+* Disable Postgresql when the JDBC overwrite is used
+* Make SonarQube Server available as an Container App on the Azure Marketplace
 
 ## [2025.5.0]
 * Upgrade SonarQube Server to to 2025.5.0